+++ Frontend-v1-Original/stores/constants/queryKeys.ts
@@ -32,3 +32,4 @@ export const VEST_VOTES = "vestVotes";
 export const GAUGES = "gauges";
 export const BASE_ASSET_INFO_NO_NATIVE = "baseAssetInfoNoNative";
 export const PAIRS_WITH_GAUGES_AND_VOTES = "pairsWithGaugesAndVotes";
+export const PAIRS_WITH_GAUGES_AND_BALANCES = "pairsWithGaugesAndBalances";
